+++ js/main.js
@@ -1,5 +1,5 @@
-import showScreen from "./screenRenderer";
-import intro from "./intro";
+import showScreen from "./utils/screen-renderer";
+import intro from "./screens/intro";
 
 showScreen(intro);
